+++ directives/docutils/__init__.py
@@ -1,0 +1,12 @@
+
+from docutils.parsers.rst import directives
+
+from .figure import DocFigure
+
+
+def register_doc_directives():
+    directives.register_directive("figure", DocFigure)
+
+
+
+